requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
streamlit==1.47.0
python-multipart==0.0.20
pydantic==2.11.7
pydantic-settings==2.10.1
PyMuPDF==1.26.3
pdfplumber==0.11.7
python-docx==1.2.0
sentence-transformers==5.0.0
rapidfuzz==3.13.0
lingua-language-detector==2.1.1
langdetect==1.0.9
numpy==2.2.6
scikit-learn==1.7.1
rapidocr-onnxruntime==1.4.4
Pillow==11.3.0
slowapi==0.1.9
httpx==0.28.1
python-dotenv==1.1.1
orjson==3.11.1
openai==1.98.0
google-genai==1.28.0
pytest==8.4.1
duckduckgo-search==8.1.1
